Patient Profile moved from pages module to apps module
