+++ apps/modernization-ui/src/apps/patient/profile/Summary.tsx
@@ -1,5 +1,5 @@
 import { PatientOpenInvestigationsTable } from 'apps/patient/profile/investigation/PatientOpenInvestigationsTable';
-import { TOTAL_TABLE_DATA } from '../../utils/util';
+import { TOTAL_TABLE_DATA } from 'utils/util';
 import { DocumentRequiringReview } from 'apps/patient/profile/documentsRequiringReview/DocumentsRequiringReview';
 
 type SummaryProp = {
